- renamed ES host var + S3 snapshot repo as config - hg* start/end as integer

+++ src/dataload/sources/wellderly/wellderly_upload.py
@@ -37,10 +37,10 @@ class WellderlyFactoryUploader(uploader.DummySourceUploader,SnpeffPostUpdateUplo
                     "hg19": {
                         "properties": {
                             "start": {
-                                "type": "long"
+                                "type": "integer"
                             },
                             "end": {
-                                "type": "long"
+                                "type": "integer"
                             }
                         }
                     },
